ICC issue: cast both sides of sign-bit-changing expressions to uint
QSGNode::DirtyForceUpdate is 0x8000, so when you shift it left by 16
positions, it becomes the sign bit. Either the result of that shift
is int or sn->dirtyState is.

 qsgbatchrenderer.cpp(302): error #68: integer conversion resulted in a change of sign

+++ src/quick/scenegraph/coreapi/qsgbatchrenderer.cpp
@@ -299,7 +299,7 @@ void Updater::updateStates(QSGNode *n)
             qDebug() << " - transforms have changed";
         if (sn->dirtyState & (QSGNode::DirtyOpacity << 16))
             qDebug() << " - opacity has changed";
-        if (sn->dirtyState & (QSGNode::DirtyForceUpdate << 16))
+        if (uint(sn->dirtyState) & uint(QSGNode::DirtyForceUpdate << 16))
             qDebug() << " - forceupdate";
     }
